aoc2022/src/test/java/dev/madfist/aoc2022/Day07Test.java
package dev.madfist.aoc2022;

import org.junit.jupiter.api.Test;

import java.util.Arrays;
import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;

public class Day07Test {
  List<String> example = Arrays.asList(
    "$ cd /",
    "$ ls",
    "dir a",
    "14848514 b.txt",
    "8504156 c.dat",
    "dir d",
    "$ cd a",
    "$ ls",
    "dir e",
    "29116 f",
    "2557 g",
    "62596 h.lst",
    "$ cd e",
    "$ ls",
    "584 i",
    "$ cd ..",
    "$ cd ..",
    "$ cd d",
    "$ ls",
    "4060174 j",
    "8033020 d.log",
    "5626152 d.ext",
    "7214296 k");

  @Test
  public void example1() {
    var day = new Day07();
    assertEquals("95437", day.solveFirst(example));
  }

  @Test
  public void example2() {
    var day = new Day07();
    assertEquals("24933642", day.solveSecond(example));
  }
}
